+++ strawberry/fastapi/router.py
@@ -70,7 +70,7 @@ class GraphQLRouter(APIRouter):
             custom_context: Optional[CustomContext],
             background_tasks: BackgroundTasks,
             connection: HTTPConnection,
-            response: Response = None,
+            response: Response = None,  # type: ignore
         ) -> MergedContext:
             request = cast(Union[Request, WebSocket], connection)
             if isinstance(custom_context, BaseContext):